# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for דרך השמש wholesale quoting system
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  barcode: string;
  description: string;
  /** Weight of content in kg */
  weightKg: number;
  /** Price per kilogram in ILS */
  pricePerKg: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  barcode: string;
  /** @minLength 1 */
  description: string;
  /** @minimum 0 */
  weightKg: number;
  /** @minimum 0 */
  pricePerKg: number;
  notes?: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  barcode?: string;
  /** @minLength 1 */
  description?: string;
  /** @minimum 0 */
  weightKg?: number;
  /** @minimum 0 */
  pricePerKg?: number;
  notes?: string;
}

export interface ProductStats {
  totalProducts: number;
  avgPricePerKg: number;
  totalWeightKg: number;
}

export interface QuoteItem {
  id: number;
  productId: number;
  barcode: string;
  description: string;
  weightKg: number;
  pricePerKg: number;
  quantity: number;
  totalPrice: number;
}

export interface QuoteItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface Quote {
  id: number;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  date: string;
  totalAmount: number;
  itemCount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface QuoteDetail {
  id: number;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  date: string;
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  items: QuoteItem[];
}

export interface QuoteInput {
  /** @minLength 1 */
  customerName: string;
  customerPhone?: string;
  date: string;
  notes?: string;
  items: QuoteItemInput[];
}

export interface QuotesSummary {
  totalQuotes: number;
  totalRevenue: number;
  recentQuotes: Quote[];
}