# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for דרך השמש wholesale quoting system
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog management
  - name: quotes
    description: Wholesale price quotes

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/stats:
    get:
      operationId: getProductStats
      tags: [products]
      summary: Get product catalog stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductStats"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quotes
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a new quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"

  /quotes/summary:
    get:
      operationId: getQuotesSummary
      tags: [quotes]
      summary: Get recent quotes summary and totals
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotesSummary"

  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a single quote with its items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required: [id, barcode, description, weightKg, pricePerKg, createdAt]
      properties:
        id:
          type: integer
        barcode:
          type: string
        description:
          type: string
        weightKg:
          type: number
          description: Weight of content in kg
        pricePerKg:
          type: number
          description: Price per kilogram in ILS
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    ProductInput:
      type: object
      required: [barcode, description, weightKg, pricePerKg]
      properties:
        barcode:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        weightKg:
          type: number
          minimum: 0
        pricePerKg:
          type: number
          minimum: 0
        notes:
          type: string

    ProductUpdate:
      type: object
      properties:
        barcode:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        weightKg:
          type: number
          minimum: 0
        pricePerKg:
          type: number
          minimum: 0
        notes:
          type: string

    ProductStats:
      type: object
      required: [totalProducts, avgPricePerKg, totalWeightKg]
      properties:
        totalProducts:
          type: integer
        avgPricePerKg:
          type: number
        totalWeightKg:
          type: number

    QuoteItem:
      type: object
      required: [id, productId, barcode, description, weightKg, pricePerKg, quantity, totalPrice]
      properties:
        id:
          type: integer
        productId:
          type: integer
        barcode:
          type: string
        description:
          type: string
        weightKg:
          type: number
        pricePerKg:
          type: number
        quantity:
          type: integer
        totalPrice:
          type: number

    QuoteItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1

    Quote:
      type: object
      required: [id, customerName, date, totalAmount, itemCount, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        date:
          type: string
          format: date
        totalAmount:
          type: number
        itemCount:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    QuoteDetail:
      type: object
      required: [id, customerName, date, totalAmount, items, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        date:
          type: string
          format: date
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"

    QuoteInput:
      type: object
      required: [customerName, date, items]
      properties:
        customerName:
          type: string
          minLength: 1
        customerPhone:
          type: string
        date:
          type: string
          format: date
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItemInput"

    QuotesSummary:
      type: object
      required: [totalQuotes, totalRevenue, recentQuotes]
      properties:
        totalQuotes:
          type: integer
        totalRevenue:
          type: number
        recentQuotes:
          type: array
          items:
            $ref: "#/components/schemas/Quote"