# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for דרך השמש wholesale quoting system
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  barcode: string;
  description: string;
  /** Weight of content in kg */
  weightKg: number;
  /** Price per kilogram in ILS */
  pricePerKg: number;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}