# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for דרך השמש wholesale quoting system
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  barcode?: string;
  /** @minLength 1 */
  description?: string;
  /** @minimum 0 */
  weightKg?: number;
  /** @minimum 0 */
  pricePerKg?: number;
  notes?: string;
}