# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for דרך השמש wholesale quoting system
 * OpenAPI spec version: 0.1.0
 */

export interface Quote {
  id: number;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  date: Date;
  totalAmount: number;
  itemCount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}