# lib/api-zod/src/generated/types/quoteDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for דרך השמש wholesale quoting system
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteItem } from './quoteItem';

export interface QuoteDetail {
  id: number;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  date: Date;
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  items: QuoteItem[];
}