# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for דרך השמש wholesale quoting system
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteItemInput } from './quoteItemInput';

export interface QuoteInput {
  /** @minLength 1 */
  customerName: string;
  customerPhone?: string;
  date: Date;
  notes?: string;
  items: QuoteItemInput[];
}