# lib/api-zod/src/generated/types/quoteItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for דרך השמש wholesale quoting system
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteItem {
  id: number;
  productId: number;
  barcode: string;
  description: string;
  weightKg: number;
  pricePerKg: number;
  quantity: number;
  totalPrice: number;
}