# lib/api-zod/src/generated/types/quoteItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for דרך השמש wholesale quoting system
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}