# lib/api-zod/src/generated/types/quotesSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for דרך השמש wholesale quoting system
 * OpenAPI spec version: 0.1.0
 */
import type { Quote } from './quote';

export interface QuotesSummary {
  totalQuotes: number;
  totalRevenue: number;
  recentQuotes: Quote[];
}